backend/requirements.txt
# Django and REST Framework
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.10

# Authentication and Security
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Vector Database
scikit-learn==1.5.0
transformers==4.45.0
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
faiss-cpu==1.8.0  # For vector indexing

# Utilities
Pillow==10.4.0  # For image processing
python-magic==0.4.27  # For file type detection
